moved tidytcells logging logic to a new iggytop.tidytcells.logging submodule
