+++ src/iggytop/tidytcells/__init__.py
@@ -1,0 +1,1 @@
+# Make tidytcells a package for import
